moving hook to shared hooks file
